Move shared tenant command validation to Application layer
